requirements.txt
aiogram==2.25.1
sqlalchemy==2.0.23
aiosqlite==0.20.0
python-dotenv==1.0.0
aiohttp==3.9.0